Improve support of of non-default XLA layout in IPU PjRt client. (#46)
Previously, the IPU PjRt client was just raising an error when called
with an input with non-default XLA layout (i.e. not C-contiguous).

This PR is improving the input stream callback implementation to add
a transpose plan in this case. A warning is still issued as the
performance cost of re-layout can be non-negligible.

Note: a factory method of uninitialized buffers with custom XLA shape
has been added to the XLA Python bindings to help testing.