Improve IPU PjRt client asynchronous dispatch performance. (#48)
In the IPU PjRt client, `Poplar::Engine::run` is being called in a
dedicated compute thread to minimize overhead. This PR is improving the
performance of this asynchronous dispatch loop by moving the cleanup
phase (i.e. deleted large state data structure) to a dedicated thread.

This removes 40-50us of overhead in the compute thread.